Unit Recognition
Tiles now correctly understand if they do or do not have a unit on them using a boolean and an object reference. Boolean gets cleared when unit is moved out of a tile (object ref cannot be made null again).
EDIT: Actually, they're still a little buggy in regards to replication. Not sure what is going wrong, need to look into it.